requirements.txt
bleach==2.1
cached-property==1.3.1
certifi==2017.7.27.1
chardet==3.0.4
cssselect==1.0.1
future==0.16.0
html5lib==1.0b10
idna==2.6
lxml==4.0.0
nltk==3.2.5
readability-lxml==0.6.2
requests==2.18.4
six==1.11.0
textblob==0.13.0
url-normalize==1.3.3
urllib3==1.22
webencodings==0.5.1
pytz==2017.3
